added ProjectTableRow component
